requirements.txt
# Nexus CRM
# Salesforce Replacement on Google Cloud Platform - Enterprise Edition

# Web Framework
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
starlette>=0.27.0,<1.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.28.0,<1.0.0
alembic>=1.12.0,<2.0.0

# Auth & Security
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
PyJWT[crypto]>=2.8.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Redis & Caching
redis>=5.0.0,<6.0.0

# XML Security
defusedxml>=0.7.0,<1.0.0

# GCP Services
google-cloud-secret-manager>=0.2.0
google-cloud-sqlconnector>=0.8.0
google-cloud-aiplatform>=1.0.0
google-cloud-logging>=3.8.0
google-cloud-pubsub>=2.18.0
google-cloud-bigquery>=3.14.0

# MCP
mcp>=1.0.0

# External APIs
httpx>=0.25.0,<1.0.0
aiohttp>=3.9.0,<4.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
python-dateutil>=2.8.2,<3.0.0
email-validator>=2.0.0,<3.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<6.0.0

# Dev
ruff>=0.1.0
mypy>=1.7.0
